requirements.txt
django==5.0.1
channels==3.0.4
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf_yasg==1.21.7
pillow==10.2.0
python-dotenv==1.0.0
requests==2.28.1
mysqlclient==2.2.1
SQLAlchemy==2.0.25
django-unfold

tronpy==0.4.0